// React/glisemikreact/src/models/ISingleFood.tsx
export interface IFood {
  createdDate?: number;
  modifiedDate?: number;
  createdBy?: string;
  modifiedBy?: string;
  gid?: number;
  cid?: number;
  name?: string;
  glycemicindex?: number;
  image?: string;
  source?: string;
  enabled?: boolean;
  detail?: string;
  url?: string;
  category?: string;
  amount?: number;
}
export type CartItemType = {
  createdDate?: number;
  modifiedDate?: number;
  createdBy?: string;
  modifiedBy?: string;
  gid?: number;
  cid?: number;
  name?: string;
  glycemicindex?: number;
  image?: string;
  source?: string;
  enabled?: boolean;
  detail?: string;
  url?: string;
  category?: string;
  amount?: number;
};